Reorg drivers (#379)
* Fix

* Use full path for parent

* Update configs

* Add the default keyword for the default instrument

* Update spectrum

* Update instruments

* Reorg drivers

* Fix laser controler

* Reorg gpio

* Move tests into dedicated instrument

* Clean up

* Move spi tests

* Move ci test script

* Move dac_router.hpp

* Rename lib -> context

* Add test_context to CI

* Fix

* Clean up

* Clean up

* Update circle.yml

* Move speed_test into test_context

* Clean up

* Merge ext_build

* Restore split_parent required for external builds

* Use sha of instrument repo for versioning

* Fix path
