+++ circle.yml
@@ -18,4 +18,4 @@ dependencies:
 
 test:
   pre:
-    - bash scripts/tests/ci_tests.sh
+    - bash scripts/ci_tests.sh
