--- drivers/laser/laser.cpp
@@ -1,21 +1,0 @@
-/// (c) Koheron
-
-#include "laser.hpp"
-
-void Laser::reset()
-{
-    xadc.set_channel(LASER_POWER_CHANNEL, LASER_CURRENT_CHANNEL);
-    xadc.enable_averaging();
-    xadc.set_averaging(256);
-    gpio.set_as_output(LASER_ENABLE_PIN, 2);
-    stop_laser();
-    set_laser_current(0.0);
-}
-
-void Laser::set_laser_current(float current)
-{
-    float current_;
-    current > MAX_LASER_CURRENT ? current_ = MAX_LASER_CURRENT : current_ = current;    
-    uint32_t pwm = pwm_from_current(current_);
-    cfg.write<reg::pwm3>(pwm);
-}
