--- drivers/lib/README.md
@@ -1,3 +1,0 @@
-### Drivers libraries
-
-Base libraries for instrument drivers development.
