+++ instruments/pulse_generator/config.yml
@@ -1,6 +1,6 @@
 ---
 instrument: pulse_generator
-parent: default
+parent: instruments/default
 board: red-pitaya
 
 cores:
