+++ instruments/test_context/run_test.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -e
+
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+HOST=$HOST py.test -v $DIR/test_memory.py
+
